move Feed type to common types
